requirements.txt
apispec==6.6.1
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==3.0.1
marshmallow==3.22.0
packaging==24.1
pycparser==2.22
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
webargs==8.6.0
Werkzeug==3.0.4
